[fix] Set default value for maxZoom and limit maxZoom option #188
The issue of empty responses from the tile server when zooming in beyond the supported zoom level has been addressed by setting the default maxZoom level to 12. This change ensures that tiles are displayed up to the supported zoom level.

If a different default maxZoom level is desired, it can be modified by adjusting the value of .                                              Closes #188