feat: Add node clustering in map view
This PR introduces a new feature that enables node clustering in the map view. This helps to improve the readability of the map when there are a large number of nodes in a small area.

This PR also includes a new example that demonstrates how to use the new feature.